feat(tests): introduce a common wait_until helper
